pyproject.toml
[tool.poetry]
name = "sdss-cerebro"
version = "0.1.1-alpha.0"
description = "A library to control Finger Lakes Instrumentation cameras."
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/cerebro"
repository = "https://github.com/sdss/cerebro"
documentation = "https://sdss-cerebro.readthedocs.io/en/latest/"
keywords = ["astronomy", "software"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "cerebro" }
]
include = [
    {path = "LICENSE.md"}
]

[tool.poetry.scripts]
cerebro = "cerebro.__main__:main"

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.17.4"
sdsstools = ">=0.3.0"
Sphinx = {version=">=3.0.0", optional=true}
influxdb-client = "^1.9.0"
click = "^7.1.2"
sdss-clu = {path='../clu'}
ntplib = "^0.3.4"
daemonocle = "^1.0.2"
click-default-group = "^1.2.2"
nest_asyncio = "^1.4.0"

[tool.poetry.dev-dependencies]
ipython = ">=7.9.0"
matplotlib = ">=3.1.1"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
rstcheck = "^3.3.1"
pytest-cov = ">=2.8.1"
pytest-sugar = ">=0.9.2"
Sphinx = ">=3.0.0"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"

[tool.poetry.extras]
docs = ["sphinx"]

[tool.isort]
line_length = 79
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = ["cerebro"]
known_sdss = ["sdsstools", "drift", "clu"]
# multi_line_output = false
balanced_wrapping = true
include_trailing_comma = false
lines_after_imports = 2
use_parentheses = true

[tool.pytest.ini_options]
addopts = "--cov cerebro --cov-report xml --cov-report html --cov-report term"

[tool.coverage.run]
branch = true
include = ["cerebro/*"]
omit = [
    "*/__init__.py",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
]

[build-system]
requires = ["poetry-core>=1.0.0a9"]
build-backend = "poetry.core.masonry.api"
